// client/safecycle-client/src/app/core/model/amenity.model.ts
export interface AmenityModel {
  type: string,
  id: number,
  lat: number,
  lon: number,
  tags: any
}